feat(OutputView): 당첨 통계내역 조회하는 기능 구현
